+++ CHANGELOG.md
@@ -5,7 +5,10 @@
 TBD
 
 - Adds the ability to delete Requests by Plugin on the System Admin page wihtin the plugin sub-menu. These are local deletes only
+- Adds Is Alive check for Entry Points, if sub process is killed externally then it is restarted
 - Add error handling if Subscriber Request has exception at creation
+- Updates request validation to support allow_any_kwargs on commands
+- Upgraded Brewtils version to [3.21.0](https://github.com/beer-garden/brewtils/releases/tag/3.21.0)
 
 # 3.19.1